+++ sqlstore/src/main/resources/db/migration/V8202__ext_inst_id_to_secondary_id.sql
@@ -1,1 +1,0 @@
-ALTER TABLE SampleTissue CHANGE COLUMN externalInstituteIdentifier secondaryIdentifier VARCHAR(255) DEFAULT NULL;